removed binary and updated file name to avoid confusion
